Move action execution related code to campaigns package
