+++ cmd/src/actions_scope_query.go
@@ -9,6 +9,7 @@ import (
 	"os"
 
 	"github.com/pkg/errors"
+	"github.com/sourcegraph/src-cli/internal/campaigns"
 )
 
 func init() {
@@ -52,7 +53,7 @@ Examples:
 			return err
 		}
 
-		var action Action
+		var action campaigns.Action
 		if err := jsonxUnmarshal(string(actionFile), &action); err != nil {
 			return errors.Wrap(err, "invalid JSON action file")
 		}
@@ -67,7 +68,7 @@ Examples:
 			}
 		}
 
-		logger := newActionLogger(*verbose, false)
+		logger := campaigns.NewActionLogger(*verbose, false)
 		repos, err := actionRepos(ctx, action.ScopeQuery, *includeUnsupportedFlag, logger)
 		if err != nil {
 			return err
